+++ CHANGES.rst
@@ -6,3 +6,7 @@ Changelog
 
 - Initial release.
   [instification]
+
+- Add configurable option to set values on all content that would match the 
+  executing rule.
+  [instification]